# [ROCm] revert cat operator performance work-around